chore(desktop): sunset legacy announcements (CDN json) system
