+++ .changeset/three-rabbits-brush.md
@@ -1,0 +1,7 @@
+---
+"ledger-live-desktop": minor
+"@ledgerhq/live-common": minor
+"@ledgerhq/live-env": minor
+---
+
+Sunset legacy announcements (CDN json) system from desktop app to gain app.json performance
