+++ libs/ledger-live-common/.unimportedrc.json
@@ -249,8 +249,6 @@
     "src/mock/account.ts",
     "src/mock/helpers.ts",
     "src/network-troubleshooting/index.ts",
-    "src/notifications/AnnouncementProvider/index.tsx",
-    "src/notifications/AnnouncementProvider/types.ts",
     "src/notifications/ServiceStatusProvider/index.tsx",
     "src/notifications/ServiceStatusProvider/types.ts",
     "src/notifications/ToastProvider/index.tsx",
@@ -494,7 +492,6 @@
     "src/modularDrawer/utils/index.ts",
     "src/modularDrawer/utils/sortByBalance.tsx",
     "src/modularDrawer/utils/type.ts",
-    "src/notifications/AnnouncementProvider/helpers.ts",
     "src/rxjs/operators/throwIf.ts",
     "src/utils/addressUtils.ts",
     "src/walletSync/getEnvironmentParams.ts"
