--- libs/ledger-live-common/src/notifications/AnnouncementProvider/api/index.ts
@@ -1,8 +1,0 @@
-import type { AnnouncementsApi } from "../types";
-import prodApi from "./api";
-
-const api: AnnouncementsApi = {
-  fetchAnnouncements: () => prodApi.fetchAnnouncements(),
-};
-
-export default api;
